Update 11/30
Comments added. Huge problem: it seems the compiler wil not run more than 5 tasks including the main, we cannot have parameters in the tasks, line 18's { has to be in the same line as it because otherwise the computer says there is another copy of main(), and it says there is a problem with the main() method because it is metioned elsewhere... which it isn't... if someone can answer all these questions please do.