Working Mobile Goal
changed moveWheels() to updateWheels() because we use moveWheels in auto. Switched lowering the carrier to a gradual decline in power. Fixed the arm method to remove power variable. Fixed carrier method to have 2 motors instead of 1. Changed the potentiometer in updateArm() for testing. I forgot to only add a couple things at a time.