Changes made at 2/10 Competition
Changes made to the autonomous section at the Milllville copetition on 2/10/18. auto2 was made to only pick up the mobile goal and park, but it doesn't quite work if I remember correctly. It shouldn't matter though because auto works, but relatively unconsistently. Will tweak auto soon. Ultrasonic sensors not quite working yet. Will be resolved soon. There are lots of unnecessary comments in auto but since they are potential ideas and auto is not yet finalized I did not remove them. stopArm was added so stopping the arm motors doesn't have to be written over and over again. Might be unnecessary though. moveWheels has been fixed so it goes in a straight line. Slight tweaks to speeds and times have been made throughout.